src/requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.10.0
attrs==25.1.0
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
distlib==0.4.0
Django==5.2.8
django-cors-headers==4.9.0
django-elasticsearch-dsl==8.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
elasticsearch==8.13.2
elasticsearch-dsl==8.13.1
elastic-transport==8.13.1
filelock==3.19.1
frozenlist==1.5.0
googletrans==3.1.0a0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
libretranslatepy==2.1.1
lxml==6.0.2
magic-filter==1.0.12
multidict==6.1.0
packaging==25.0
pillow==12.0.0
platformdirs==4.4.0
propcache==0.2.1
psycopg2-binary==2.9.11
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
translate==3.6.1
translation==1.0.5
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.34.0
whitenoise==6.11.0
yarl==1.18.3
